Fix bugs related to public room browsing
