+++ src/core/hle/service/http/http_c.cpp
@@ -14,6 +14,7 @@
 #include "common/file_util.h"
 #include "common/scope_exit.h"
 #include "common/string_util.h"
+#include "common/web_util.h"
 #include "core/core.h"
 #include "core/file_sys/archive_ncch.h"
 #include "core/file_sys/file_backend.h"
@@ -111,53 +112,6 @@ constexpr Result ErrorInvalidPostDataEncoding = // 0xD8A0A035
 constexpr Result ErrorIncompatibleSendPostData = // 0xD8A0A036
     Result(ErrCodes::IncompatibleSendPostData, ErrorModule::HTTP, ErrorSummary::InvalidState,
            ErrorLevel::Permanent);
-
-// Splits URL into its components. Example: https://citra-emu.org:443/index.html
-// is_https: true; host: citra-emu.org; port: 443; path: /index.html
-static URLInfo SplitUrl(const std::string& url) {
-    const std::string prefix = "://";
-    constexpr int default_http_port = 80;
-    constexpr int default_https_port = 443;
-
-    std::string host;
-    int port = -1;
-    std::string path;
-
-    const auto scheme_end = url.find(prefix);
-    const auto prefix_end = scheme_end == std::string::npos ? 0 : scheme_end + prefix.length();
-    bool is_https = scheme_end != std::string::npos && url.starts_with("https");
-    const auto path_index = url.find("/", prefix_end);
-
-    if (path_index == std::string::npos) {
-        // If no path is specified after the host, set it to "/"
-        host = url.substr(prefix_end);
-        path = "/";
-    } else {
-        host = url.substr(prefix_end, path_index - prefix_end);
-        path = url.substr(path_index);
-    }
-
-    const auto port_start = host.find(":");
-    if (port_start != std::string::npos) {
-        std::string port_str = host.substr(port_start + 1);
-        host = host.substr(0, port_start);
-        char* p_end = nullptr;
-        port = std::strtol(port_str.c_str(), &p_end, 10);
-        if (*p_end) {
-            port = -1;
-        }
-    }
-
-    if (port == -1) {
-        port = is_https ? default_https_port : default_http_port;
-    }
-    return URLInfo{
-        .is_https = is_https,
-        .host = host,
-        .port = port,
-        .path = path,
-    };
-}
 
 static std::size_t WriteHeaders(httplib::Stream& stream,
                                 std::span<const Context::RequestHeader> headers) {
@@ -291,7 +245,7 @@ void Context::MakeRequest() {
         {RequestMethod::PutEmpty, "PUT"},
     };
 
-    URLInfo url_info = SplitUrl(url);
+    Common::URLInfo url_info = Common::SplitUrl(url);
 
     httplib::Request request;
     std::vector<Context::RequestHeader> pending_headers;
@@ -375,7 +329,7 @@ void Context::MakeRequest() {
     }
 }
 
-void Context::MakeRequestNonSSL(httplib::Request& request, const URLInfo& url_info,
+void Context::MakeRequestNonSSL(httplib::Request& request, const Common::URLInfo& url_info,
                                 std::vector<Context::RequestHeader>& pending_headers) {
     httplib::Error error{-1};
     std::unique_ptr<httplib::Client> client =
@@ -395,7 +349,7 @@ void Context::MakeRequestNonSSL(httplib::Request& request, const URLInfo& url_in
     }
 }
 
-void Context::MakeRequestSSL(httplib::Request& request, const URLInfo& url_info,
+void Context::MakeRequestSSL(httplib::Request& request, const Common::URLInfo& url_info,
                              std::vector<Context::RequestHeader>& pending_headers) {
     httplib::Error error{-1};
     X509* cert = nullptr;
